tdf#156560 Fix assertion failure editing RTL header
Clicking on an RTL header previously lead to assertion failure, but now
it is checked if the SwFrame pointer is null. If so, an empty vector is
returned.

It should be noted that the rendering of the rtl header in the
attachment 188689 is still wrong as the text and image are misplaced.

A test is added to the patch, and can be run using:

 make CppunitTest_sw_uiwriter8 CPPUNIT_TEST_NAME=testTdf156560

Change-Id: I9e8f66895b4ef9369300c9d7c03d22a2085a1d71
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155159
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>