tdf#152307 sw: layout: invalidate more frames when footer grows
While on page 12:
SwTabFrame::MakeAll() is called on 523 while it's on page 12; there is
one invalid pos text frame 492 with a fly somewhere before it;
PrepareMake() of 523 formats prevs 492 then 493 which MoveFwd() taking
523 with it.

While on page 13:
TabFrame 523 is valid, and the footer 6651 never formatted (0 height).
Formatting the footer in SwHeadFootFrame::FormatSize() invalidates the
body 1031, immediately calls Calc() from SwFrame::MakePos(), where
~SwLayNotify() -> SwLayoutFrame::ChgLowersProp() invalidates
SectionFrame 1034.

Then SectionFrame 1034 is formatted, which via
SwSectionFrame::CheckClipping() -> SwLayoutFrame::ChgLowersProp()
invalidates only the last lower frame, because it checks the position of
the frame and this frame still has a position on a previous page (it
moved from page 12) and isn't even invalid yet.

So in case there are invalid frames, the positions of the frames
following these cannot be trusted to be used to optimize invalidations
in SwLayoutFrame::ChgLowersProp().

(aside: it seems odd to format the body text before the footer text, but
 in this case doing it differently wouldn't have helped because the
 problem was already caused on a previous page)

(regression from commit b9ef71476fd70bc13f50ebe80390e0730d1b7afb)

Change-Id: I23b35c09af3a373d0913d931a2ba59d45fadf2c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155196
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>