tdf#156260 sw floattable: avoid moving text from the last anchor to its precede
The 5 page long bugdoc had a "dt" in the last paragraph, expanding that
to an actual dummy text with F3 resulted in anchor text on both page 4
and page 5, while the expected behavior is that text only wraps on the
last page, i.e. page 5.

What happened is that the (text) "offset" of the text frame anchors on
pages 1..5 are meant to be all 0 (so the actual anchor text is only on
the last page), but in practice the anchor text frame on page 5 had an
offset of 1123, which breaks the invariant that in case a fly frame is
split to N pages, then only the last matching anchor text frame has
text. (If it has too much text, then a next page with just anchor text
is fine, but 0..N-1 pages should have no anchor text.)

Fix the problem by SwTextFrame::AdjustFollow_(): in case it would reach
the end for the "has follow, but no follow's follow" case, then still
avoid setting the offset of the follow in case we know this is text
frame has a non-last split fly anchored in it.

With this, the render result matches in no overlaps and also what Word
does in a similar edit session. (No F3 there, but can paste similar text
at the document end.)

Change-Id: I6c8d4f825d9dc356bb2a3a87a2f9c2e6529ce533
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155209
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins