tdf#156473 a11y: Use actual row/col index/count in grid control header
The row header bar of a grid control is a table with
just one column, containing a row header cell for
each of the rows in the table holding the actual data.
Likewise, a column header bar just has one row.

This was previously incorrectly reported on the a11y layer,
as could e.g. be seen in Accerciser with the sample macro
from tdf#156473.

Also make the `AccessibleGridControlCell` ctor pass the
`_nColumnRowId` as what it actually is (a row index
in case of a row header cell, a column index in case
of a column header cell) instead of always as a row
index, and then having special handling for that
elsewhere again
(s. `AccessibleGridControlCell::getAccessibleName`).

Also move an assert from
`AccessibleGridControlCell::getAccessibleName` to
the ctor and check the indices there.

With this in place, the row and column header bars
for the example from tdf#156473 now report the correct
amount of rows and columns, and the row/column header
cells report the correct row and column indices.

Change-Id: I29f71ac46b6d841e26d68ca01bd05ba9412aed13
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155224
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>