tdf#156525 Save as > HTML loses drawing object as invalid gif
regression from
    commit 1c7cbd685633d44eac554629572f3401c450f855
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Sun May 7 16:56:21 2023 +0200
    use AlphaMask for variables when calling GetAlphaMask

where after my change, the code is now calling
    Bitmap::Replace(AlphaMask,...)
instead of
    Bitmap::Replace(Bitmap,...)
and those two methods do quite different things.

However, we have to
(*) restore Bitmap::Replace(Bitmap,...) which was removed in
    commit 8270eb5d5600cc84dbf5f0e339f90c4519ef88bb
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Fri May 19 13:35:31 2023 +0200
    loplugin:unusedmethods
(*) restore BitmapWriteAccess::SetPaletteEntryCount which was removed in
    commit 74cd0d0b281f8df75612bfb600df2eae62c4d21d
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Thu Jun 29 13:53:30 2023 +0200
    loplugin:unusedmethods
(*) Invert the mask/alpha layer, since after
    commit 81994cb2b8b32453a92bcb011830fcb884f22ff3
    Author: Noel Grandin <noelgrandin@gmail.com>
    Date:   Fri Apr 16 20:33:10 2021 +0200
    Convert internal vcl bitmap formats transparency->alpha (II)
we are dealing with real alpha, and not transparency.

Also add an assert in GIFWriter::WriteAccess, since it is a logic error
to get here and have the image in the wrong format.

Change-Id: I0e09b3ca82af0bd5b58d80e0a6eac4c7bdf7c48e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155254
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>