tdf#145743 sw: don't delete empty page with ColLocked section
In MakeFrames(), the new frames are inserted into a section frame,
before an existing frame; the existing frame is MoveFwd() to the next
page, so the call to CheckPageDescs() finds the page frame empty and
deletes it, along with the section frame.

Check the mbColLocked flag set by MakeFrames() to prevent it.

(regression from commit b9ef71476fd70bc13f50ebe80390e0730d1b7afb)

Change-Id: I4e02a85314adaf07794fd6b5447bd78d39597b8d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155266
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>