tdf#62603 Fix find/replace to not extend font attributes
This change modifies the logic of the core replace code
to now not replace the 1st character and extend its
attributes, but instead to actually replace characters
to keep respective formatting.  Additional checks
are made for whether the replacement was shorter or
longer than the found text, and to handle trimming
or appending the final portions as needed.

Change-Id: I03a5645898e55ad386bacc2766af9b244a97bd21
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155274
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>