gtk3 a11y tests: Increase DBus timeout
We apparently get recurrent spurious timeout failures in the CI
related to DBus timeout, so increase that value.

The chosen value is a bit arbitrary, but is below the maximum
recommended by the documentation (even though the recommendation hardly
applies to our use case), and more than 3 times larger than the
default.  That will hopefully be enough.

Change-Id: I244f711f46353009743015d00b303aeb87332936
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155310
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>