sw floattable, crashtesting: fix PDF export of tdf73201-1.docx
The direct cause is that in in the SwTextFrameBreak ctor, m_pFrame has
IsInSct() as true, but FindSctFrame() then doesn't find anything, which
should not happen.

The bigger problem is that currently the fly split code assumes that
frames grow from (logical) top to bottom, so in case the vertical
orientation is bottom, and we have to grow from bottom to top, that
needs re-positioning the fly, but the fly split code has no support for
that.

Fix the crash by denying this case in SwFlyFrame::IsFlySplitAllowed()
for now, the document's table is not supposed to split in the first
place.

In case later there is a real need to split such flys, then more
research is needed how that re-positioning works in detail for non-split
flys and then we have to adapt the fly split code accordingly.

Change-Id: I56cf074a3d740ab716f7de232fbd131d17544d9e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155339
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins