tdf#147938 sw: Undo of table to text uses ContentIdxStore wrong
ContentIdxStore must be created for every SplitContentNode() with its
content index and used there; SwNodes::UndoTableToText() creates it once
per text node which is rather unhelpful and in this case results in a
redline with incorrect range, starting in one table cell and ending in
another.

This problem exists since initial CVS import.

Change-Id: I8a20b13b42384658bf869c5202ccbad2213c12f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155344
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>