tdf#147938 sw_redlinehide: fix undo of table to text
The problem is that when the text node is split in
SwNodes::UndoTableToText(), the 2nd one retains the merge flag NonFirst,
not the 1st one, and it ends up in the 2nd table cell, so it's skipped
when creating the layout frames (and an assert in InsertCnt_()
segfaults).

SwContentNode::DelFrames() should probably reset the merge flags of
nodes when it deletes the merged frame.

(regression from commit 32902f66e7749b2d06d13f50416be5323a0c0ea9)

Change-Id: I7b3b1ca873bde831481c676780200a3206ae667c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155345
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>