Related tdf#156508 fix rendering of the About dialog .svg
After 0e7d4ea64f84bdc62c10692dcd97c3a32e72e16e, white polygons are
drawn in the LibreOffice about .svg in both the About dialog and with
the .svg in a document.

Creating the mask with COL_WHITE instead of COL_BLACK appears to fix
this without causing tdf#156508 to reoccur.

Change-Id: I886451245126c2b7f936e30275965148bf029627
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155362
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>