ofz#57493 Timeout
the oss-fuzzer has gotten itself into that nasty state where it
can't reproduce the original problem (because its fixed) but keeps
finding other unreproducible ones that don't allow the fixed one
to get closed. Disable this part of the fuzzer for a while so the
original issue can close and then we can reenable and new findings will
be recorded as a new fresh bug.

Change-Id: Ie73c5c50a850a0ac6ea51345f6478f5750f8248e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155385
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>