tdf#156551 tdf#150606 sw: layout: only invalidate SwTabFrame if it...
... wants to move back.

In this document, the table 1709 splits across 3 columns of the section
1702 with 1 row each, and when its follows are invalidated in
CalcContent(), the later SwTabFrame::MakeAll() of the 1st follow joins
the 2nd follow and invalidates the size of the section frame, causing a
loop.

So only invalidate if that actually looks necessary.

(regression from commit 59987d3c77ec7dbf59fbea9f47cc226f4e8903f9)

Change-Id: I360c8f697a7666a19a08d8ebcabbcfa3a7cdf844
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155452
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>