tdf#147666 sw: layout: no IsShortCut() if fly was moved to page
The problem was that the fly's anchor frame was moved to page 2, and
then IsShortCut() returns true for the anchor frame, which means only
the anchor frame itself is formatted, but not its fly.

Thus the fly remains invalid and positioned on page 1 when its rectangle
is passed to SwViewShell::VisPortChgd(), and only later it is properly
positioned.

(regression from commit eb85de8e6b61fb3fcb6c03ae0145f7fe5478bccf)

Change-Id: I9e0554bd0d9751b47a11a3917575d382139ee93a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155461
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>