tdf#156679 wina11y: Implement accHitTest via UNO equivalent
Instead of manually iterating over the children
and checking whether the given position is in the location
that they're in in `CMAccessible::accHitTest`, use
`XAccessibleComponent::XAccessibleComponent`, which
provides exactly the functionality that's needed.
(This is similar to what the Qt-based VCL plugins
on Linux do, s. `QtAccessibleWidget::childAt`.)

This also drops the need to limit this to objects
that have at most a certain amount of children
for performance reasons (previously 256) and thus
makes this work e.g. also to identify a Calc cell
that the mouse pointer is currently over while previously
the document was returned, as could be seen also in
NVDA's Python console:

1) start NVDA
2) hover over a Calc cell
3) press NVDA+Ctrl+Z to capture snapshot variables
4) check what NVDA identifies as the mouse object
   via the Python Console:

Before:

    >>> mouse
    <NVDAObjects.IAccessible.IAccessible object at 0x00FB9910>
    >>> mouse.name
    'Untitled 1 - LibreOfficeDev Spreadsheets'
    >>> mouse.role
    <Role.DOCUMENT: 52>

With this change in place:

    >>> mouse
    <NVDAObjects.Dynamic_SymphonyIATableCellEditableTextWithAutoSelectDetectionIAccessible object at 0x0774DD10>
    >>> mouse.name
    >>> mouse.role
    <Role.TABLECELL: 29>

The cell's text still isn't announced by NVDA even
with mouse tracking enabled, but that's another issue.

Change-Id: Ib821020cef6303ab786c4c3fc3ccd917398214f1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155491
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>