tdf#142489 Do not show the table toolbar when Notebookbar is active
When the Notebookbar UI is being used, all the commands are already accessible via the "Table" tab in the notebookbar, so showing the toolbar duplicates the commands on the screen.

Therefore, this patch disables showing this toolbar when the Notebookbar UI is being used. Obviously, it can be enabled manually via View - Toolbars.

Change-Id: I115effbd94f61df518ff90411a45431fefa9e844
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155473
Tested-by: Jenkins
Tested-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>