SCriptForge (SFDatabases) manage database credentials
When not embedded, databases usually
require user and password to grant
access to scripts.

When the access was refused, so far the
execution was stopped brutally with an
error message provided by LO internal
code.
This was made visible during some tests
with a MySql database server.

The code to be reviewed is located in the
creation of the "database" service. This
service has the sungularity that it can
be called from several other service
instances.

As from this commit unsuccessful tentatives
to login to a database from code are
rejected with a clean SF error message.

The different scenarios to create a
database service have been (re)tested.

Change-Id: I695d108242872d27671688af76916784908aeebe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155507
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins