tdf#147938 sw_redlinehide: move fix into SwNodes::UndoTableOfText()
Reproducing tdf#151866 on the .odt attachment, it becomes clear that the
fix for tdf#147938 is in the wrong place.

There a node has the merge flag None although it is the last node of a
merge, because CheckParaRedlineMerge() first sets the node flags and
then calls DelFrames(), which now resets the flags.

In case there's another caller of DelFrames() that needs this, it's not
obvious, can be added later.

(regression from commit 5b55ccfd384d3c0c11e05f22bd7e956b4bc7e43a)

Change-Id: Iba0cb83260ecb18a4b96ded20a95b2b20d696fc4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155516
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>