tdf#156209 Font size increases when copying from cell in edit mode
Regression from  0558926c2f9201a12b4c46efc36b8a4080af4d46
"use more officecfg for SvtMiscOptions".

However.... even reverting the above commit does not fix the problem.
This is because there is an underlying latent bug, where depending on the order of initialisation of various and when they fire events, we can end up calling the alternative path in ScInputHandler::ImplCreateEditEngine which creates a pool via EditEngine::CreatePool(), which ends up having the "wrong" default map unit. And then that pool sticks around even after we have a document, and causes the reported bug.

Solve it by simply not initialising the pool until we have an associated document, which is fine, since we don't actually need the pool until then.


Change-Id: If012c82df225e59dca828556bfca84c1948b5887
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155560
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>