Related tdf#156629 force snapshot of alpha mask
On macOS, with Skia/Metal with a Retina display (i.e. 2.0 window
scale), the alpha mask gets upscaled for the font color and the
character background color icons in Writer's properties sidebar
with most icon sets. The Breeze and Colibre icon sets running in
light mode are the most obvious cases.

This bug appears to be caused by asynchronous rendering of the
returned image. So, we force a copy of the alpha mask in case it
changes before the image is actually drawn.

Change-Id: Ie3aa3b26acaee4bd9173e6b356137980c070c7c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155562
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>