Related tdf#156630 and tdf#156629 force snapshot of alpha mask
On macOS, with Skia/Metal or Skia/Raster with a Retina display
(i.e. 2.0 window scale), the alpha mask gets upscaled in certain
cases.

This bug appears to be caused by pending scaling of an existing
SkImage in the bitmap parameter. So, force the SkiaSalBitmap to
handle its pending scaling.

This change consolidates the three duplicate fixes added in the
following commits into the AlphaMask(const Bitmap&) constructor.

  commit 12fd870113a663dde5ceb38c61f1986a34095d0e
  commit ce2d9f5dd4b6a26847c4779bce4866d969ff4400

From my light testing, the bitmap's SkImage is not scaled except
when running macOS with a Retina display. The only exception is
that this fix will be triggered on all platforms when exporting
to PDF images with an alpha mask.

Change-Id: Iea5afd55aac984ca606b2b4b44e457d81d76fac0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155568
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>