tdf#155495 sc: fix cursor positioning editing centered/right-aligned cells
It was not always possible to place the cursor in edited text
by mouse clicking in case of centered and right-aligned cell
content, i.e. when the edited content has been shown before
the original (not edited) cell position.

Now it's possible to place the cursor by mouse clicking before
the left border of the not edited cell (instead of selecting other
cells covered by the edited content).

Change-Id: I23c59ca3ef67d223181a7e3dfd71b84f43fca32b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155579
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>