qt: Implement pinch/zoom gesture handling
Subscribe to receive pinch gestures by calling
`QWidget::grabGesture(Qt::PinchGesture)` and
forward that as the corresponding `SalGestureZoomEvent`.

From looking at what values the gtk implementation uses,
`1 + pPinchGesture->totalScaleFactor()` seems to be a
proper choice for the
`SalGestureZoomEvent::mfScaleDelta` value.

With this in place, zooming in and out of a Writer
document with a two-finger zoom gesture works for
me with the Qt-based VCL plugins in Writer on my laptop
that has a touch screen.
(It seems to be a bit more smooth with qt5/kf5 than
with qt6, where doing several zoom gestures after
each other sometimes requires to wait a bit before
doing another gesture in order for that one to be
processed properly.)

Qt documentation for gestures: [1]

Corresponding commit for gtk VCL plugins:

    commit f2bd19f6720239db228cd4388be8e928505c51b6
    Author: Povilas Kanapickas <povilas@radix.lt>
    Date:   Thu Aug 25 00:18:30 2022 +0300

        vcl: implement touchpad zoom gesture support on gtk backend

[1] https://doc.qt.io/qt-6/gestures-overview.html

Change-Id: I288943d923463fee44314969648e03dca84c483f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155649
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>