tdf#156724 tdf#156722 tdf#156745 sw: layout: partially remove IsPaintLocked()
Having the layout algorithm dependend on IsPaintLocked() is very
problematic; it is typically set when the layout is invoked from
SwViewShell code, as happens several times when loading a document in
response to window resize events etc., but not for idle formatting or
from SwXTextDocument::getRendererCount(), hence these bugs only
reproduce with soffice --convert-to pdf, not via UI.

For tdf#156724 the problem is that the table is split, during formatting
of the split row a new footnote is created on page 1, and this reduces
the space, so splitting fails and is never attempted again.

When the document is loaded from UI, when the table is split the
footnote already exists and so splitting succeeds; it was created by
a call from SwLayAction::FormatLayout() of the cell frame.

It turns out that when the condition is removed completely, testUXTSOREL
will take 5 minutes instead of 5 seconds, which seems excessive; the
problem there appears to be that a text frame in a columned section
moves forward and backward; plausibly columned section content should be
formatted by ::CalcContent() only.

(reportedly regression from commit c605283ad6785dea762feab5fdffd9d27e75c292 and commit
7e8b4756d95057f069467b34e7849f9354856578)

Change-Id: I9ed73588efeec654a769eee8aa825186bd51e059
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155672
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit dc5991ccc8f7d8e86c1e04be9095dda5c08b4763)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155691
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>