tdf#155470 DOCX export: fix RTL numbers changed to LTR
When using numerals in "context" mode in "Tools > Options", in
"Language Settings > Complext Text Layout > General Options > Numerals",
the actual display of the numbers is dependant on the direction and
the language of those numbers.

Previously, when exporting the DOCX in LO the RTL direction was lost,
and because of that, the numerals were shown incorrectly after saving
in LibreOffice, and loading again.

This patch adds <w:rtl/> where <w:lang> contains an RTL language.
For example, before <w:lang w:bidi="fa-IR"/> inside <w:rPr>, it always
adds <w:rtl/>.

It is worth noting that this is not always the case, as there are
situations that <w:rtl/> is not present in <w:rPr> with an RTL language
like <w:lang w:bidi="fa-IR"/> and <w:rtl/> should not be in the output.
This patch does not handle that case, because currently loading such a
text is not done correctly in LO, and the LTR numbers inside RTL
environment are shown just like RTL numbers. To fix it, first loading
and displaying of that should be fixed.

A test is added to check the export of the sample file, which can be
invoked with:

 make CppunitTest_sw_ooxmlexport CPPUNIT_TEST_NAME=testTdf155470

Change-Id: I8fc54916fcd3295010c5c81ccb632bc6082ffaa2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155721
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>