sw: PDF/UA export: add /Placement/Block to Note
PDF 1.7 defines Note as an ILSE (inline), but that is rather odd and
definitely not how footnotes work in Writer.

PAC 3 complains about "Possibly inappropriate use of a "Note" structure
element" and "Possibly inappropriate use of a "P" structure element"
which is rather opaque but experimenting shows the warnings disappear
when the Note has a /Placement/Block attribute.

PDF 2.0 (ISO-32000-2:2020 Annex M) even removes the Note ILSE and
introduces FENote as a BLSE to replace it.

Change-Id: I46ee32c88d4be495bf79549280e2fa68d7da091c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155722
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>