tdf#156682 sw floattable: fix missing del of master anchor para por on split
Pressing "enter" at the end of B1 in the bugdoc resulted in a split
floating table, which had the anchor text duplicated (both on page 1 and
page2).

What happened is that the master anchor frame didn't have its paragraph
portion cleared, the follow anchor frame had a correct frame offset (of
0).

Fix the problem in SwTextFrame::FormatEmpty(), similar to what commit
00b9b33334791079c2dc26b1ed4c123450cabf7d (sw: call FormatEmpty() in
SwTextFrame::Format() for split fly masters, 2023-02-09), did; here what
we want to avoid is setting some properties on the paragraph preventing
the deletion of non-last anchor paragraph portions.

As a side effect, this also fixes the hang from the bugreport, after
pressing "enter" 5 times at the end of B1.

Change-Id: I16e6d786f02ad503a0fdff465ecf7716c1c5f8b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155732
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins